Add files via upload
Origional V.1 Upload for AppiForm; a simple form that logs items by dumping information into .json. If an avaliable .json file is not avaliable, the system will create one.
This is ran on Python 3 and is only runnable (as a current) by Python 3.
Modifies are not allowed unless specific permission is given by Ness Leickons (creator)
A lock may be set in order the system not be modifiable.